Move title-chrome helpers from app/helpers to Views::FullPageBase
Title-family setters (`add_page_title`, `add_show_title`,
`add_edit_title`, `add_new_title`, `add_index_title`,
`add_owner_naming`) plus `add_query_filters` move from
`app/helpers/header/{title,filters}_helper.rb` to
`Views::FullPageBase` as instance methods. Both helper files are
deleted; the corresponding `register_value_helper` lines in
`Views::Base` go with them.

Three classification fixes surface from running the suite without
the helper-on-Views::Base safety net:

- `Views::Controllers::Descriptions::Versions::Show`,
  `Locations::Descriptions::Merges::New`,
  `Names::Descriptions::Merges::New` were inheriting `Views::Base`
  despite being action views — promoted to `Views::FullPageBase`.
- `VisualGroupTable` was a sub-partial that called `add_show_title`;
  moved the call up to its two action-view callers
  (`VisualGroups::Index` and `VisualModels::Show`), so the table
  stays a pure rendering sub-partial.

One Phlex-semantics gotcha caught: `render(SomeView.new)` in Phlex
2.x emits to the buffer and returns nil, unlike ActionView's
`render` which returned a string. `add_show_title` / `add_edit_title`
/ `add_query_filters` wrap their `render` calls in `capture` so the
HTML string lands in the `content_for` slot.

The deleted `Header::TitleHelperTest` is replaced by
`test/views/full_page_base/title_chrome_test.rb` exercising the same
utility methods (`title_tag_contents`, `document_title_for`,
`show_document_title`, `edit_document_title`) via `send`. The
side-effect setters (`add_*_title` family) are covered end-to-end by
the existing controller tests for each show / edit / new / index
page that calls them; both `Names::` and `Locations::Descriptions::
VersionsControllerTest` grow display assertions on the `<title>`
element + body class.
