+++ app/views/controllers/visual_models/visual_group_table.rb
@@ -18,11 +18,15 @@ module Views::Controllers::VisualModels
     prop :visual_groups, _Array(VisualGroup)
 
     def view_template
-      add_show_title(@visual_model)
       render_top_nav
       h3 { plain("Visual Groups") }
       render_table
     end
+
+    # Both callers (`VisualGroups::Index`, `VisualModels::Show`)
+    # render this table as the entire page body. Each calls
+    # `add_show_title(@visual_model)` itself so the page-chrome
+    # call sits on the action view, not on the sub-partial.
 
     private
