+++ app/views/layouts/header/page_title.rb
@@ -9,9 +9,8 @@
 #   - right (show-only, non-project): interest-icons strip and the
 #     prev/index/next pager.
 #
-# Both `show_page_edit_icons` and `show_page_interest_icons` (formerly
-# helper-module methods in `Header::TitleHelper`) are inlined here as
-# private methods — this was the only caller of either.
+# `show_page_edit_icons` / `show_page_interest_icons` are private
+# methods on this view — the only caller of either.
 module Views::Layouts
   class Header::PageTitle < Views::Base
     SHOW_TITLE_CLASSES =
