+++ test/controllers/locations/descriptions/versions_controller_test.rb
@@ -12,7 +12,15 @@ module Locations::Descriptions
       desc.reload
       new_versions = desc.versions.length
       assert(new_versions > old_versions)
+
       get(:show, params: { id: desc.id })
+
+      # `Views::Controllers::Descriptions::Versions::Show` extends
+      # `Views::FullPageBase`, so the Application layout fires —
+      # doctype + chrome + page-title slot populated.
+      assert_response(:success)
+      assert_select("html > head > title", count: 1)
+      assert_select("body.versions__show")
       assert_select("#description_details_and_alts")
     end
   end
