+++ test/controllers/names/descriptions/versions_controller_test.rb
@@ -12,8 +12,16 @@ module Names::Descriptions
       desc.reload
       new_versions = desc.versions.length
       assert(new_versions > old_versions)
+
       get(:show, params: { id: desc.id })
-      assert_select("#description_details_and_alts")
+
+      # `Views::Controllers::Descriptions::Versions::Show` extends
+      # `Views::FullPageBase`, so the full Application layout fires:
+      # doctype + layout chrome + the page-title slot populated by
+      # `add_page_title` from `view_template`.
+      assert_response(:success)
+      assert_select("html > head > title", count: 1)
+      assert_select("body.versions__show")
       assert_select("#description_details_and_alts")
     end
   end
